+++ src/main/java/fzmm/zailer/me/utils/SkinPart.java
@@ -2,15 +2,15 @@ package fzmm.zailer.me.utils;
 
 import fzmm.zailer.me.utils.position.PosI;
 
-public record SkinPart(int x, int y, int hatX, int hatY) {
+public record SkinPart(byte x, byte y, byte hatX, byte hatY) {
 
     // https://imgur.com/3LlJdua
-    public static final SkinPart RIGHT_LEG = new SkinPart(16, 48, 0, 48);
-    public static final SkinPart RIGHT_ARM = new SkinPart(32, 48, 48, 48);
-    public static final SkinPart LEFT_LEG = new SkinPart(0, 16, 0, 32);
-    public static final SkinPart LEFT_ARM = new SkinPart(40, 16, 40, 32);
-    public static final SkinPart BODY = new SkinPart(16, 16, 16, 32);
-    public static final SkinPart HEAD = new SkinPart(0, 0, 32, 0);
+    public static final SkinPart RIGHT_LEG = new SkinPart((byte) 16, (byte) 48, (byte) 0, (byte) 48);
+    public static final SkinPart RIGHT_ARM = new SkinPart((byte) 32, (byte) 48, (byte) 48, (byte) 48);
+    public static final SkinPart LEFT_LEG = new SkinPart((byte) 0, (byte) 16, (byte) 0, (byte) 32);
+    public static final SkinPart LEFT_ARM = new SkinPart((byte) 40, (byte) 16, (byte) 40, (byte) 32);
+    public static final SkinPart BODY = new SkinPart((byte) 16, (byte) 16, (byte) 16, (byte) 32);
+    public static final SkinPart HEAD = new SkinPart((byte) 0, (byte) 0, (byte) 32, (byte) 0);
 
     public PosI getNormalLayer() {
         return new PosI(this.x, this.y);
